Test automatic wrapping of model manager
The docs say that any custom model manager will be automatically wrapped
with the money_manager (at least as long as it's named 'objects').
(https://github.com/django-money/django-money#important-note-on-model-managers)

I'm having some issue where it seems like that doesn't happen, though.
Trying to add a test case to show that it doesn't work as the
documentation explains it.

Am I misunderstanding something?